+++ packages/@expo/metro-config/src/serializer/fork/__tests__/serializer-test-utils.ts
@@ -2,7 +2,6 @@ import assert from 'assert';
 
 import { microBundle, projectRoot } from './mini-metro';
 import { reconcileTransformSerializerPlugin } from '../../reconcileTransformSerializerPlugin';
-import { treeShakeSerializer } from '../../treeShakeSerializerPlugin';
 import {
   SerialAsset,
   SerializerConfigOptions,
@@ -66,12 +65,12 @@ export async function serializeSplitAsync(
   });
 }
 
-// Serialize to a split bundle
-export async function serializeShakingAsync(
+export async function serializeOptimizeAsync(
   fs: Record<string, string>,
   options: {
     isReactServer?: boolean;
     treeshake?: boolean;
+    optimize?: boolean;
     splitChunks?: boolean;
     minify?: boolean;
   } = {}
@@ -83,13 +82,19 @@ export async function serializeShakingAsync(
         platform: 'web',
         dev: false,
         output: 'static',
-        treeshake: true,
+        treeshake: false,
+        optimize: true,
         splitChunks: true,
         minify: false,
         inlineRequires: true,
         ...options,
       },
     },
-    [treeShakeSerializer, reconcileTransformSerializerPlugin]
+    [reconcileTransformSerializerPlugin]
   );
+}
+
+export function expectImports(graph, name: string) {
+  if (!graph.dependencies.has(name)) throw new Error(`Module not found: ${name}`);
+  return expect([...graph.dependencies.get(name).dependencies.values()]);
 }
